value::operator datetime: fix clang warning

+++ src/value.cpp
@@ -2823,7 +2823,7 @@ namespace tds {
 
                 memcpy(&n, d.data(), 3);
 
-                return datetime{num_to_ymd(n - jan1900), time_t(0)};
+                return datetime{num_to_ymd((int)(n - jan1900)), time_t(0)};
             }
 
             case sql_type::TIME: {
